test: restore private unit test modules
