+++ crates/djls-server/Cargo.toml
@@ -26,8 +26,6 @@ tracing-appender = { workspace = true }
 tracing-subscriber = { workspace = true }
 
 [dev-dependencies]
-djls-testing = { workspace = true }
-
 salsa = { workspace = true }
 tempfile = { workspace = true }
